Fix bug in De Sterck code, to calculate fev correcly.
De Sterck made a mistake in his code when calculating function
evaluations. The Poblano toolbox linesearch reports the number of
function evaluations, but does so in the sixth output value.
De Sterck incorrectly used the fifth output, which reports the "info"
flag from the More-Thuente line search. This flag is often 1, even when
fev > 1.